Upload Project
This is a small project. Its main purpose is to generate blank papers -
applications for civil cases in different Bulgarian courts. Reads
information from file "info.xlsx". After collecting all the information,
it creates "person" objects that hold all the necessary information.
Then he reads a blank and fills the information at specific places in it
and generates blank forms - applications that are ready for printing and
distribution to the courts.